# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for Stash-box metadata database",
        "title": "Stash-box REST API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Stash App",
            "url": "https://github.com/stashapp/stash-box"
        },
        "license": {
            "name": "AGPL-3.0",
            "url": "https://www.gnu.org/licenses/agpl-3.0.en.html"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/edits/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth SessionAuth": []
                    }
                ],
                "tags": [
                    "edits"
                ],
                "summary": "Get edit by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Edit ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_models.Edit"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/performers": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth SessionAuth": []
                    }
                ],
                "tags": [
                    "performers"
                ],
                "summary": "List performers with filters",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 25,
                        "description": "Items per page",
                        "name": "per_page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by name",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by gender",
                        "name": "gender",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by ethnicity",
                        "name": "ethnicity",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by country",
                        "name": "country",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by eye color",
                        "name": "eye_color",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by hair color",
                        "name": "hair_color",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Minimum height in cm",
                        "name": "height_min",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Maximum height in cm",
                        "name": "height_max",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Minimum age",
                        "name": "age_min",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Maximum age",
                        "name": "age_max",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort field (name, birthdate, scene_count, created_at, updated_at)",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort direction (asc, desc)",
                        "name": "direction",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.PaginatedResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_stashapp_stash-box_internal_models.Performer"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth SessionAuth": []
                    }
                ],
                "tags": [
                    "performers"
                ],
                "summary": "Create a new performer",
                "parameters": [
                    {
                        "description": "Performer data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_models.PerformerCreateInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_models.Performer"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/performers/check-existing": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth SessionAuth": []
                    }
                ],
                "tags": [
                    "performers"
                ],
                "summary": "Check for existing performers",
                "parameters": [
                    {
                        "description": "Query criteria",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_models.QueryExistingPerformerInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_stashapp_stash-box_internal_models.Performer"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/performers/edits/{editId}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth SessionAuth": []
                    }
                ],
                "tags": [
                    "performers"
                ],
                "summary": "Update a pending performer edit",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Edit ID (UUID)",
                        "name": "editId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated performer edit data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api_rest_handlers_performer.UpdatePerformerEditRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_models.Edit"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/performers/search": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth SessionAuth": []
                    }
                ],
                "tags": [
                    "performers"
                ],
                "summary": "Search performers by term",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search term",
                        "name": "term",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 25,
                        "description": "Result limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_stashapp_stash-box_internal_models.Performer"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_gofrs_uuid.NullUUID": {
            "type": "object",
            "properties": {
                "uuid": {
                    "type": "string"
                },
                "valid": {
                    "type": "boolean"
                }
            }
        },
        "github_com_stashapp_stash-box_internal_api_rest_handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_stashapp_stash-box_internal_api_rest_handlers.PaginatedResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "page": {
                    "type": "integer"
                },
                "per_page": {
                    "type": "integer"
                },
                "total_count": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "github_com_stashapp_stash-box_internal_api_rest_handlers.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {}
            }
        },
...
        "github_com_stashapp_stash-box_internal_models.Edit": {
            "type": "object",
            "properties": {
                "applied": {
                    "type": "boolean"
                },
                "bot": {
                    "type": "boolean"
                },
                "closed_at": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "id": {
                    "type": "string"
                },
                "operation": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "target_type": {
                    "type": "string"
                },
                "update_count": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "$ref": "#/definitions/github_com_gofrs_uuid.NullUUID"
                },
                "votes": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_stashapp_stash-box_internal_models.QueryExistingPerformerInput": {
            "type": "object",
            "properties": {
                "disambiguation": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "urls": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "API Key authentication",
            "type": "apiKey",
            "name": "ApiKey",
            "in": "header"
        },
        "SessionAuth": {
            "description": "Session cookie authentication",
            "type": "apiKey",
            "name": "stashbox",
            "in": "cookie"
        }
    }
}